+++ game/internal/surplus_pool.go
@@ -25,6 +25,6 @@ const (
 //盈余池
 var SurplusPool float64 = 0
 
-func SetPackageTaxM(packageT uint16, tax uint8) {
+func SetPackageTaxM(packageT uint16, tax float64) {
 	packageTax[packageT] = tax
 }
